requirements_evaluation.txt
# Requirements for model evaluation
tensorflow>=2.16.0,<2.21.0
numpy>=1.26.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
h5py>=3.10.0